requirements.txt
opencv-python>=4.10
numpy>=1.26
pyyaml>=6.0
python-dotenv>=1.0
requests>=2.32
ultralytics>=8.3
